move math functions to __functions
