+++ adaptivetesting/math/estimators/__ml_estimation.py
@@ -2,7 +2,7 @@ from typing import List, Tuple
 import jax.numpy as np
 from ...models.__test_item import TestItem
 from ...services.__estimator_interface import IEstimator
-from .functions.__estimators import maximize_likelihood_function
+from .__functions.__estimators import maximize_likelihood_function
 
 
 class MLEstimator(IEstimator):
